fixed all compiler warnings

+++ src/driver/disk/ahci/ahci.cpp
@@ -57,7 +57,7 @@ AHCI::AHCI(pci::pci_device_header_t* pci_base_address) {
 	renderer::global_font_renderer->printf("AHCI driver instance initialized.\n");
 	this->pci_base_address = pci_base_address;
 
-	AHCI::ABAR = (HBA_memory*)((pci::pci_header_0_t*)pci_base_address)->BAR5;
+	AHCI::ABAR = (HBA_memory*)(uint64_t)((pci::pci_header_0_t*)pci_base_address)->BAR5;
 
 	g_page_table_manager.map_memory(AHCI::ABAR, AHCI::ABAR);
